fix calendar disabled end time days & times
