+++ ui/src/components/dialogs/TournamentConfirmation.tsx
@@ -22,14 +22,14 @@ import {
 } from "@/lib/utils";
 import { getTokenLogoUrl, getTokenSymbol } from "@/lib/tokensMeta";
 import { useEkuboPrices } from "@/hooks/useEkuboPrices";
-import { Settings, Token } from "@/generated/models.gen";
+import { Token } from "@/generated/models.gen";
 import { useGameEndpoints } from "@/dojo/hooks/useGameEndpoints";
 import { useDojoStore } from "@/dojo/hooks/useDojoStore";
-import { SettingsDetails } from "@/generated/models.gen";
 import { useMemo } from "react";
 import { useDojo } from "@/context/dojo";
 // import { calculateTotalValue } from "@/lib/utils/formatting";
 import { useGetGameSettings } from "@/dojo/hooks/useSqlQueries";
+import { mergeGameSettings } from "@/lib/utils/formatting";
 
 interface TournamentConfirmationProps {
   formData: TournamentFormData;
@@ -68,47 +68,7 @@ const TournamentConfirmation = ({
     (token) => token.address === formData?.gatingOptions?.token
   );
 
-  const settingsEntities = [...settingsDetails, ...settings];
-
-  console.log(settingsEntities);
-
-  const mergedGameSettings = useMemo(() => {
-    if (!settingsEntities) return {};
-
-    return settingsEntities.reduce(
-      (acc, entity) => {
-        const details = entity.models[gameNamespace ?? ""]
-          .SettingsDetails as SettingsDetails;
-        const settings = entity.models[gameNamespace ?? ""]
-          .Settings as Settings;
-        const detailsId = details.id.toString();
-
-        // If this details ID doesn't exist yet, create it
-        if (!acc[detailsId]) {
-          acc[detailsId] = {
-            ...details,
-            hasSettings: false,
-            settings: [],
-          };
-        }
-
-        // If we have settings, add them to the array and set hasSettings to true
-        if (settings) {
-          acc[detailsId].settings.push(settings);
-          acc[detailsId].hasSettings = true;
-        }
-
-        return acc;
-      },
-      {} as Record<
-        string,
-        SettingsDetails & {
-          hasSettings: boolean;
-          settings: Settings[];
-        }
-      >
-    );
-  }, [settingsEntities, formData.game]);
+  const mergedGameSettings = mergeGameSettings(settingsDetails, settings);
 
   const hasBonusPrizes =
     formData.bonusPrizes && formData.bonusPrizes.length > 0;
